Working except for pixelation
The refresh button is finally doing everything it was intended to do; however, when you press the button, the user's pixelation slider value is not retrieved. This means it works properly if they do not change the slider between times that the wallpaper is refreshed, but if the slider is changed the images will not reflect it unless you change to a different Wallpaper Engine wallpaper then switch back.